Revert to playing title screen tune in BASIC
While slightly hacky in places, this actually feels like a fairly solid
basis for a pseudo-fixed frame rate game. Next thing to do would
probably be to merge in/redo the optimisations I experimented with
before so we can do each cycle faster and be more likely to want to
busy-wait.