Hide all objects when collected, not just prisms
This is hacky because when the object type isn't known, the bug in
u_subroutine wasn't so easy to work around and I decided to patch it. I
should at this point probably bite the bullet and remove the need for
the BASIC to poke &5600/&5700 as well.

Moderately tested and does seem to work for all four prisms, the health
pickup and the fleece. (I have a feeling the fleece disappeared anyway,
but this doesn't seem to have broken it anyway.)