Protect machine code above HIMEM
Before HIMEM was at &5800 and it just happened that there was a bit of
room for the BASIC stack after the end of slot_wrap_behaviour_table, but
nothing except luck/discipline stopped the BASIC stack growing down over
machine code/data that mattered.

With this change the BASIC code can use as much stack as it likes
without corrupting anything, subject to actually having enough memory
free of course.