Bugfix, tweaks
I wasn't restoring X before the bcc. This doesn't make things work
properly, but it probably makes them work better.

I *was* able to get myself stuck in a wall in the first room with this
code, but I'm fairly sure I had passed "through" the enemy as this
happened, which this code doesn't even pretend to deal with. This is
nowhere near tested enough to say it "works" as long as enemies don't
get involved yet; I just played it once for a minute or so.