Could this finally be the bug fix?
The hang test caught that ri_y was not always in the range 0-2 on
entering s_subroutine. The old code would treat that consistently (off
the top of my head, not sure what as) but now we do more checks on ri_y
at more points invalid values trigger inconsistent behaviour.